+++ backend/aws/__init__.py
@@ -1,0 +1,5 @@
+"""All code for interacting with AWS"""
+
+from .dynamo import Dynamo
+from .eventbridge import EventBridge
+from .secretmanager import SecretManager
