+++ backend/newsletter/package.json
@@ -3,5 +3,8 @@
   "version": "1.0.0",
   "description": "Example of creating a function that runs as a cron job using the serverless `schedule` event",
   "author": "",
-  "license": "MIT"
+  "license": "MIT",
+  "devDependencies": {
+    "serverless-python-requirements": "^5.2.2"
+  }
 }
