+++ backend/play_counter/handler.py
@@ -4,12 +4,13 @@ from os import environ
 
 from botocore.exceptions import ClientError
 
-from aws.aws import AWS
+from aws import Dynamo, EventBridge, SecretManager
+from aws.models.lambda_func import LambdaAction
 from clients.spotify import SpotifyClient
-from telemetry.logging import LambdaAction, log_execution, log_failure, logger
+from telemetry.logging import logger
 
 
-def reschedule(aws_client: AWS, listening_now: bool):
+def reschedule(eb_client: EventBridge, listening_now: bool):
     """
     Reschedules the EventBridge rule responsible for triggering this lambda
 
@@ -19,7 +20,7 @@ def reschedule(aws_client: AWS, listening_now: bool):
     listening_now: bool
     """
     eb_rule = environ["EVENT_BRIDGE_RULE"]
-    current_rule_schedule = aws_client.get_rule_schedule(name=eb_rule)
+    current_rule_schedule = eb_client.get_rule_schedule(name=eb_rule)
 
     if (
         time.fromisoformat("02:00:00")
@@ -29,7 +30,7 @@ def reschedule(aws_client: AWS, listening_now: bool):
         logger.info(
             "Rescheduling Execution", current=current_rule_schedule, new="rate(1 hour)"
         )
-        aws_client.update_event_rule(name=eb_rule, rate=1, period="hour")
+        eb_client.update_event_rule(name=eb_rule, rate=1, period="hour")
         return
 
     if not listening_now and (current_rule_schedule != "rate(5 minutes)"):
@@ -38,7 +39,7 @@ def reschedule(aws_client: AWS, listening_now: bool):
             current=current_rule_schedule,
             new="rate(5 minutes)",
         )
-        aws_client.update_event_rule(name=eb_rule, rate=5, period="minutes")
+        eb_client.update_event_rule(name=eb_rule, rate=5, period="minutes")
         return
 
     if listening_now and (current_rule_schedule == "rate(5 minutes)"):
@@ -47,7 +48,7 @@ def reschedule(aws_client: AWS, listening_now: bool):
             current=current_rule_schedule,
             new="rate(1 minute)",
         )
-        aws_client.update_event_rule(name=eb_rule, rate=1, period="minute")
+        eb_client.update_event_rule(name=eb_rule, rate=1, period="minute")
         return
 
     logger.info("Continuing at Existing Schedule", current=current_rule_schedule)
@@ -62,14 +63,15 @@ def run(_, context):
     context:
         the lambda function metadata
     """
-    log_execution(context.function_name, LambdaAction.TRIGGERED)
-    aws_client = AWS()
-    sp_client = SpotifyClient(aws_client=aws_client)
+    logger.info(f"Function {LambdaAction.TRIGGERED}", function=context.function_name)
+    sp_client = SpotifyClient(dynamo_client=Dynamo(), secret_manager=SecretManager())
 
     try:
         current_song = sp_client.get_current_song()
-        reschedule(aws_client=aws_client, listening_now=(current_song is not None))
+        reschedule(eb_client=EventBridge(), listening_now=(current_song is not None))
     except ClientError as ex:
-        log_failure(context.function_name, ex)
+        logger.exception(
+            f"Function {LambdaAction.FAILED}", funcion=context.function_name, error=ex
+        )
 
-    log_execution(context.function_name, LambdaAction.COMPLETED)
+    logger.info(f"Function {LambdaAction.COMPLETED}", function=context.function_name)
