+++ backend/telemetry/logging.py
@@ -1,21 +1,8 @@
 """Code for logging"""
 
-from enum import Enum
-
 from structlog import configure
 from structlog.processors import JSONRenderer, TimeStamper
 from structlog.stdlib import BoundLogger, add_log_level, get_logger
-
-
-class LambdaAction(Enum):
-    """
-    States representing pieces of the lambda function lifecycle
-    """
-
-    TRIGGERED = 1
-    COMPLETED = 2
-    FAILED = 3
-
 
 configure(
     wrapper_class=BoundLogger,
@@ -29,26 +16,3 @@ configure(
 )
 
 logger = get_logger()
-
-
-def log_execution(func_name: str, action: LambdaAction):
-    """
-    Logs a key moment in lambda execution
-
-    func_name: str
-        the name of the lambda function
-    action: FunctionAction
-        the lambda lifecycle event to log for
-    """
-    logger.info(f"Function {action.name}", function_name=func_name)
-
-
-def log_failure(func_name: str, exception: Exception):
-    """
-    Logs the failure of a lambda function
-    """
-    logger.error(
-        f"Function {LambdaAction.FAILED.name} due to exception",
-        function_name=func_name,
-        exception=str(exception),
-    )
